+++ core/terraform-invocs.nix
@@ -1,20 +1,39 @@
-{ pkgs ? import <nixpkgs> { }, terraformConfiguration, terraformWrapper ? pkgs.terraform, prefixText ? "" }:
+{ pkgs ? import <nixpkgs> { }, terraformConfiguration, terraformWrapper ? null, prefixText ? "" }:
 let
+  wrapper =
+    # Use wrapper attrset
+    if terraformWrapper ? package then {
+      package = terraformWrapper.package;
+      extraRuntimeInputs = terraformWrapper.extraRuntimeInputs or [];
+      prefixText = terraformWrapper.prefixText or "";
+      suffixText = terraformWrapper.suffixText or "";
+    }
+    # Assume this is pkgs.terraform, pkgs.opentofu or other package derivation
+    else {
+      package = if terraformWrapper == null then pkgs.terraform else terraformWrapper;
+      extraRuntimeInputs = [];
+      prefixText = prefixText;
+      suffixText = "";
+    };
+
+  tfBinaryName = wrapper.package.meta.mainProgram;
+
   mkTfScript = name: text: pkgs.writeShellApplication {
     inherit name;
-    runtimeInputs = [ terraformWrapper ];
+    runtimeInputs = [ wrapper.package ] ++ wrapper.extraRuntimeInputs;
     text = ''
-      ${prefixText}
+      ${wrapper.prefixText}
       ln -sf ${terraformConfiguration} config.tf.json
-      terraform init
+      ${tfBinaryName} init
       ${text}
+      ${wrapper.suffixText}
     '';
   };
 in
 rec {
   scripts = {
-    apply = mkTfScript "apply" "terraform apply";
-    destroy = mkTfScript "destroy" "terraform destroy";
+    apply = mkTfScript "apply" "${tfBinaryName} apply";
+    destroy = mkTfScript "destroy" "${tfBinaryName} destroy";
   };
 
   apps = pkgs.lib.fix (self:
@@ -22,8 +41,6 @@ rec {
     // { default = self.apply; });
 
   devShells.default = pkgs.mkShell {
-    buildInputs = (builtins.attrValues scripts) ++ [ terraformWrapper ];
+    buildInputs = (builtins.attrValues scripts) ++ [ wrapper.package ];
   };
 }
-
-